modal added to the page
