+++ src/index.html
@@ -20,7 +20,7 @@
           <a class="nav-link" href="index.html">Home <span class="sr-only">(current)</span></a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="">Fattest Best Award</a>
+          <a class="nav-link" id="bearBest" href="">Fattest Bear Award</a>
         </li>
         <li class="nav-item dropdown">
           <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -57,18 +57,41 @@
         </div>
         <button type="submit" class="btn btn-primary" id="submitBear">Submit</button>
       </form>
-    </div>
-  
+    </div> 
+    <!-- Button trigger modal -->
+    <button type="button" class="btn-lg btn-warning mt-3" id="fattest" data-toggle="modal" data-target="#exampleModalCenter">
+      Fattest Bear Awards
+    </button>
 
-    <div class="row mt-5" id="bearCards">
+    <!-- Modal -->
+    <div class="modal fade" id="exampleModalCenter" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+      <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-content bg-dark">
+          <div class="modal-header text-center">
+            <h5 class="modal-title text-warning" id="exampleModalLongTitle">Fattest Bear Awards</h5>
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="modal-body">
+            ...
+          </div>
+          <div class="modal-footer">
+            <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+            <button type="button" class="btn btn-primary">Save changes</button>
+          </div>
+        </div>
+      </div>
+    </div>
+
+
+    <div class="row mt-3" id="bearCards">
       <div class="col border border-primary text-white">All the bears you are stalking today
         <div class="col-sm d-flex justify-content-around flex-wrap" id="bearContainer"></div>
       </div>
     </div>
   </div>
   <div id="app"></div>
-
-
 
   <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-Piv4xVNRyMGpqkS2by6br4gNJ7DXjqk09RmUpJ8jgGtD7zP9yug3goQfGII0yAns" crossorigin="anonymous"></script>
